--- hackasm/cmd/hackasm/main_test.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"path/filepath"
-	"testing"
-)
-
-func TestExec(t *testing.T) {
-	srcPath := "./testdata/PongL.asm"
-	tmp := t.TempDir()
-	outPath := filepath.Join(tmp, "out.hack")
-	exec(srcPath, outPath)
-}
